User password contextualization
Add the "CRYPTED_PASSWORD attribute value as encrypted password
for both "root" and "docker" users. As a safety precaution,
when this attribute is not set, disable both passwords altogether
instead of using the default "tcuser" password for "docker".

Root password is of a limited usage, because password auth
is disabled for root in sshd config.