Add HTTP_PROXY context variable
Allow HTTP proxy to be set up from the outside context.
Unfortunately, docker daemon is already running when we set up
the context, so we would either need to split the contextualization
into more stages, or restart docker after the proxy is configured.

I went for the later quick and dirty approach, sorry.